Update README.md
This Python project uses machine learning and Keras to classify images as either cats or dogs. It involves training a deep learning model on a dataset of labeled cat and dog images. The model is built using Keras, specifically a convolutional neural network (CNN) architecture. The dataset is divided into training and testing sets, with the training set used to train the model and the testing set used to evaluate its performance. The model is compiled with a loss function, optimizer, and evaluation metrics, and then trained on the training set. Finally, the model's performance is assessed on the testing set to determine its classification accuracy.